Made translation possible with larger graphs and added m1 probabilities

+++ src/include/travatar/hyper-graph.h
@@ -11,7 +11,7 @@
 
 namespace travatar {
 
-typedef short NodeId;
+typedef int NodeId;
 class HyperNode;
 class HyperGraph;
 class TranslationRule;
